pyproject.toml
[project]
name = "event-elo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "pip>=25.3",
    "streamlit-aggrid>=1.1.9",
]
